rename firestorex to modelfs
