+++ go/app/ticket/internal/modelfs/ticket_test.go
@@ -1,4 +1,4 @@
-package firestorex_test
+package modelfs_test
 
 import (
 	"testing"
@@ -7,13 +7,13 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/t-ash0410/stack-example/go/app/ticket/internal/firestorex"
+	"github.com/t-ash0410/stack-example/go/app/ticket/internal/modelfs"
 )
 
 var (
 	t2024_12_29_UTC = time.Date(2024, 12, 29, 0, 0, 0, 0, time.UTC)
 
-	baseTicket = &firestorex.Ticket{
+	baseTicket = &modelfs.Ticket{
 		TicketID:    "083c61da-b38d-4a8c-9c2d-f7ff466678b5",
 		Title:       "Some Ticket",
 		CreatedBy:   "8ea79f88-5b4b-4df6-b438-81a2ccf6b09f",
@@ -34,7 +34,7 @@ func TestTicket_Validate(t *testing.T) {
 	t.Run("Fail", func(t *testing.T) {
 		t.Parallel()
 
-		invalid := &firestorex.Ticket{
+		invalid := &modelfs.Ticket{
 			TicketID:    "", // important
 			Title:       "", // important
 			CreatedBy:   "", // important
