+++ go/app/ticket/internal/modelfs/validate.go
@@ -1,4 +1,4 @@
-package firestorex
+package modelfs
 
 import "github.com/go-playground/validator/v10"
